backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
trimesh[all]==4.1.4
numpy==1.26.4
cascadio==0.2.1
gunicorn==21.2.0
cadquery-ocp==7.9.3.1
requests==2.31.0
python-dotenv==1.0.1
